little modification on Makefile

+++ Makefile
@@ -14,7 +14,7 @@ DEV=/dev/ttyUSB0
 
 #location
 LEFT = 10.492
-BOTTOM = 52.254
+BOTTOM = 52.2407
 RIGHT = 10.5551
 TOP = 52.2801
 
@@ -26,8 +26,14 @@ FILTER2 = $(FILTER1)
 INV_FILTER2 = $(INV_FILTER1)
 
 
+compile: $(OUTPUT).gpx
+
+#nur bei parameter aufruf grafisch anzeigen
 run: $(OUTPUT).gpx
 	prune.sh $(OUTPUT).gpx
+
+#alle meist genutzten ausgaben generieren
+all: ic_pb ic_rest
 
 send: $(OUTPUT).gpx
 	#sendet die gpx datei auf das Garmin
@@ -46,22 +52,24 @@ $(OUTPUT).gpx: $(INPUT)
 
 $(INPUT):
 		wget -O "$(INPUT)" "http://www.informationfreeway.org/api/0.6/*[$(OSM_KEY)=$(OSM_VALUE)][bbox=$(LEFT),$(BOTTOM),$(RIGHT),$(TOP)]"
+
 #oder http://osmxapi.hypercube.telascience.org/api/0.6/*[amenity=shop][bbox=10.492,52.254,10.5551,52.2801]
 
 #incomplete post_boxes
 ic_pb:
-	rm post_box.gpx
+	[ -e post_box.gpx ] && rm post_box.gpx || echo ""
 	make INPUT=post_box.osm FILTER1=collection_times INV_FILTER1=yes FILTER2=operator INV_FILTER2=yes
 
 #incomplete restaurant
 ic_rest:
-	rm restaurant.gpx
+	[ -e restaurant.gpx ] && rm restaurant.gpx || echo ""
 	make INPUT=restaurant.osm OSM_VALUE=restaurant FILTER1=opening_hours INV_FILTER1=yes
-	#FILTER2=cuisine INV_FILTER2=no
 
-#complete post_boxes - isnt realy possible cause a node will be selected if filter1 or filter2 is true
+#FILTER2=cuisine INV_FILTER2=no
+
+#complete post_boxes - isnt really possible cause a node will be selected if filter1 or filter2 is true
 c_pb:
-	rm post_box.gpx
+	[ -e post_box.gpx ] && rm post_box.gpx || echo ""
 	make INPUT=post_box.osm FILTER1=collection_times INV_FILTER1=no
